travail du jour reglage d'erreurs

+++ lexer/parse_read.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_read.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ctruchot <ctruchot@student.42.fr>          +#+  +:+       +#+        */
+/*   By: aduvilla <aduvilla@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/09 16:23:28 by aduvilla          #+#    #+#             */
-/*   Updated: 2024/03/11 18:17:27 by ctruchot         ###   ########.fr       */
+/*   Updated: 2024/03/12 18:31:40 by aduvilla         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,17 +20,20 @@ t_cmd	*parse_read(char *read, char **env)
 	ft_bzero(&parse, sizeof(t_parse));
 	if (!read)
 		return (NULL);
+	if (check_read(read))
+		return (NULL);
 	if (create_token_list(&parse.lexer, read))
 		return (ft_listclear(&parse.lexer), NULL);
-	replace_argument(&parse.lexer, env);
+	if (replace_argument(&parse.lexer, env))
+		return (ft_listclear(&parse.lexer), NULL);
 	delete_quotes(&parse.lexer);
 	if (search_errors(parse.lexer))
 		return (ft_listclear(&parse.lexer), NULL);
-	create_cmd_list(&parse);
+	if (create_cmd_list(&parse))
+		return (ft_listclear(&parse.lexer), ft_cmd_lstclear(&parse.cmd), NULL);
 	print_cmd_lst(parse.cmd);
 	print_lst(parse.lexer);
 	ft_listclear(&parse.lexer);
-	ft_printf("avant clear\n");
 	// ft_cmd_lstclear(&parse.cmd);
 	return (parse.cmd);
 }
